% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-model.R
\name{export_model}
\alias{export_model}
\alias{export_keras_model}
\alias{export_keras_model.AutokerasModel}
\alias{export_autokeras_model}
\alias{export_autokeras_model.AutokerasModel}
\title{Creates and exports the Keras or AutoKeras model to the given filename.}
\usage{
export_keras_model(object, ...)

\method{export_keras_model}{AutokerasModel}(autokeras_model,
  model_file_name)

export_autokeras_model(object, ...)

\method{export_autokeras_model}{AutokerasModel}(autokeras_model,
  model_file_name)
}
\arguments{
\item{model_file_name}{Output file path to save the model. File should be
.h5 for Keras, and .pkl for AutoKeras.}
}
\description{
Creates and exports the Keras or AutoKeras model to the given filename.
}
