% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\alias{fit.AutokerasModel}
\title{Find the best neural architecture and train it.}
\usage{
fit.AutokerasModel(autokeras_model, x, y, time_limit)
}
\arguments{
\item{autokeras_model}{An AutokerasModel instance.}

\item{x}{An array containing the training data or the training data combined
with the validation data.}

\item{y}{An array containing the label of the training data. or the label of
the training data combined with the validation label.}

\item{time_limit}{The time limit for the search in seconds.}
}
\description{
Based on the given dataset, the function will find the best neural
architecture for it.
The dataset is in array format.
So they training data should be passed through x and y.
}
