% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keras_model.R
\name{get_keras_model}
\alias{get_keras_model}
\alias{get_keras_model.AutokerasModel}
\title{Return the best trained Keras architecture.}
\usage{
get_keras_model(autokeras_model)

\method{get_keras_model}{AutokerasModel}(autokeras_model)
}
\arguments{
\item{autokeras_model}{A trained AutokerasModel instance.}
}
\description{
Return the best trained Keras architecture.
}
