% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-image-regressor.R
\name{model_image_regressor}
\alias{model_image_regressor}
\title{Auto-Keras Model}
\usage{
model_image_regressor(path = NULL, verbose = TRUE, augment = TRUE,
  resume = FALSE)
}
\arguments{
\item{path}{A path to the directory to save the classifier as well as
intermediate results.}

\item{verbose}{A boolean of whether the search process will be printed to
stdout.}

\item{augment}{A boolean value indicating whether the data needs
augmentation. True by default.}

\item{resume}{A boolean. If True, the classifier will continue to previous
work saved in path. Otherwise, the classifier will start a new search.}
}
\description{
ImageRegressor class.
It is used for image regression. It searches convolutional neural network
architectures for the best configuration for the image dataset.
}
\details{
Important: The object returned by this function behaves like an R6 object,
i.e., within function calls with this object as parameter, it is most likely
that the object will be modified. Therefore it is not necessary to assign
the result of the functions to the same object.
}
