% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_structured_data_classifier.R
\name{model_structured_data_classifier}
\alias{model_structured_data_classifier}
\title{Auto-Keras Model}
\usage{
model_structured_data_classifier(
  column_names = NULL,
  column_types = NULL,
  num_classes = NULL,
  multi_label = FALSE,
  loss = NULL,
  metrics = NULL,
  name = "structured_data_classifier",
  max_trials = 100,
  directory = NULL,
  objective = "val_accuracy",
  overwrite = TRUE,
  seed = NULL
)
}
\arguments{
\item{column_names}{: A list of characters specifying the names of the
columns. The length of the list should be equal to the number of columns of
the data excluding the target column. Defaults to `NULL`. If `NULL`, it
will obtained from the header of the csv file or the `data.frame`.}

\item{column_types}{: A list of characters. The names are the column names.
The values should either be 'numerical' or 'categorical', indicating the
type of that column. Defaults to `NULL`. If not `NULL`, the `column_names`
need to be specified. If `NULL`, it will be inferred from the data.}

\item{num_classes}{: numeric. Defaults to `NULL`. If `NULL`, it will infer
from the data.}

\item{multi_label}{: logical. Defaults to `FALSE`.}

\item{loss}{: A Keras loss function. Defaults to use "binary_crossentropy" or
"categorical_crossentropy" based on the number of classes.}

\item{metrics}{: A list of Keras metrics. Defaults to use "accuracy".}

\item{name}{: character. The name of the AutoModel. Defaults to
"structured_data_classifier".}

\item{max_trials}{: numeric. The maximum number of different Keras Models to
try. The search may finish before reaching the `max_trials`. Defaults to
`100`.}

\item{directory}{: character. The path to a directory for storing the search
outputs. Defaults to `NULL`, which would create a folder with the name of
the AutoModel in the current directory.}

\item{objective}{: character. Name of model metric to minimize or maximize.
Defaults to "val_accuracy".}

\item{overwrite}{: logical. Defaults to `TRUE`. If `FALSE`, reloads an
existing project of the same name if one is found. Otherwise, overwrites
the project.}

\item{seed}{: numeric. Random seed.}
}
\description{
AutoKeras structured data classification class.
}
\details{
Important: The object returned by this function behaves like an R6 object,
i.e., within function calls with this object as parameter, it is most likely
that the object will be modified. Therefore it is not necessary to assign
the result of the functions to the same object.
}
\examples{
# Create a structured data classifier
clf <- model_structured_data_classifier()
}
