% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tabular-regressor.R
\name{model_tabular_regressor}
\alias{model_tabular_regressor}
\title{Auto-Keras Model}
\usage{
model_tabular_regressor(path = NULL)
}
\arguments{
\item{path}{A path to the directory to save the classifier as well as
intermediate results.}
}
\description{
TabularRegressor class.
}
\details{
Important: The object returned by this function behaves like an R6 object,
i.e., within function calls with this object as parameter, it is most likely
that the object will be modified. Therefore it is not necessary to assign
the result of the functions to the same object.
}
