% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-text-regressor.R
\name{model_text_regressor}
\alias{model_text_regressor}
\title{Auto-Keras Model}
\usage{
model_text_regressor(path = NULL, verbose = TRUE)
}
\arguments{
\item{path}{A path to the directory to save the classifier as well as
intermediate results.}

\item{verbose}{A boolean of whether the search process will be printed to
stdout.}
}
\description{
TextRegressor class.
}
