% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.AutokerasModel}
\title{Return predict results for the testing data.}
\usage{
\method{predict}{AutokerasModel}(autokeras_model, x_test)
}
\arguments{
\item{autokeras_model}{A trained AutokerasModel instance.}

\item{x_test}{An array containing the testing data.}
}
\description{
Return predict results for the testing data.
}
