% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\alias{evaluate.AutokerasModel}
\title{Return the accuracy score between predict value and y_test.}
\usage{
evaluate.AutokerasModel(autokeras_model, x_test, y_test)
}
\arguments{
\item{autokeras_model}{A trained AutokerasModel instance.}

\item{x_test}{An array containing the testing data.}

\item{y_test}{An array containing the testing targets.}
}
\description{
Return the accuracy score between predict value and y_test.
}
