% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_model.R
\name{export_model}
\alias{export_model}
\title{Export model}
\usage{
export_model(autokeras_model)
}
\arguments{
\item{autokeras_model}{: A trained AutokerasModel instance.}
}
\value{
keras.Model instance. The best model found during the search, loaded
  with trained weights.
}
\description{
Export the best trained Keras Model.
}
