% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_model.R
\name{export_model}
\alias{export_model}
\alias{export_keras_model}
\alias{export_keras_model.AutokerasModel}
\alias{export_autokeras_model}
\alias{export_autokeras_model.AutokerasModel}
\alias{import_autokeras_model}
\alias{import_autokeras_model.character}
\title{Export and import functions for Keras and AutoKeras models.}
\usage{
export_keras_model(autokeras_model, model_file_name)

\method{export_keras_model}{AutokerasModel}(autokeras_model, model_file_name)

export_autokeras_model(autokeras_model, model_file_name)

\method{export_autokeras_model}{AutokerasModel}(autokeras_model, model_file_name)

import_autokeras_model(model_file_name)

\method{import_autokeras_model}{character}(model_file_name)
}
\arguments{
\item{autokeras_model}{A trained AutokerasModel instance.}

\item{model_file_name}{Output/Input file path to save/load the model. File
should be .h5 for Keras, and .pkl for AutoKeras.}
}
\description{
Creates and exports the Keras or AutoKeras model to the given filename.
Or imports the previously exported AutoKeras model.
}
