% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-image-classifier.R
\name{model_image_classifier}
\alias{model_image_classifier}
\title{Auto-Keras Model}
\usage{
model_image_classifier(path = NULL, verbose = TRUE, augment = TRUE,
  resume = FALSE)
}
\arguments{
\item{path}{A path to the directory to save the classifier as well as
intermediate results.}

\item{verbose}{A boolean of whether the search process will be printed to
stdout.}

\item{augment}{A boolean value indicating whether the data needs
augmentation. True by default.}

\item{resume}{A boolean. If True, the classifier will continue to previous
work saved in path. Otherwise, the classifier will start a new search.}
}
\description{
ImageClassifier class.
It is used for image classification. It searches convolutional neural
network architectures for the best configuration for the image dataset.
}
