% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-tabular-classifier.R
\name{model_tabular_classifier}
\alias{model_tabular_classifier}
\title{Auto-Keras Model}
\usage{
model_tabular_classifier(path = NULL, verbose = TRUE)
}
\arguments{
\item{path}{A path to the directory to save the classifier as well as
intermediate results.}

\item{verbose}{A boolean of whether the search process will be printed to
stdout.}
}
\description{
TabularClassifier class.
}
