% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-text-classifier.R
\name{model_text_classifier}
\alias{model_text_classifier}
\title{Auto-Keras Model}
\usage{
model_text_classifier(path = NULL, verbose = TRUE, glove_file = NULL)
}
\arguments{
\item{path}{A path to the directory to save the classifier as well as
intermediate results.}

\item{verbose}{A boolean of whether the search process will be printed to
stdout.}

\item{glove_file}{A path where the glove.zip must be downloaded, or where it
was previously downloaded. This file is going to be used by Auto-Keras to
train the model. GloVe: Global Vectors for Word Representation.}
}
\description{
A TextClassifier class based on Google AI's BERT model.
}
\details{
Important: The object returned by this function behaves like an R6 object,
i.e., within function calls with this object as parameter, it is most likely
that the object will be modified. Therefore it is not necessary to assign
the result of the functions to the same object.
}
