% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict.AutokerasModel}
\title{Predict}
\usage{
\method{predict}{AutokerasModel}(object, x, batch_size = 32, ...)
}
\arguments{
\item{object}{: A trained AutokerasModel instance.}

\item{x}{: Any allowed types according to the input node. Testing data. Check
corresponding AutokerasModel help to note how it should be provided.}

\item{batch_size}{: numeric. Defaults to `32`.}

\item{...}{: Unused.}
}
\description{
Predict the output for a given testing data.
}
